refactor(app): extract the connection-picker wiring
Selecting a connection into the detail panel, the split-pane tab
handlers, collapsing/deleting/renaming sidebar groups, the picker search
box, favourites, drag-reorder between groups, disconnect, reconnect and
the external-link opener move to `wire::picker`.

The cluster is cut in two around the tab save/restore closures, which
stay in `main` because everything downstream builds on them.
